updated code

+++ infra/Scripts/destroy_infra.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 read -p "Do you want to use 'set -e' option (Stop when command fail)? (Y/N): " SET_E
-if [[ "$SET_E" == "Y" ]]; then
+if [[ "${SET_E^^}" == "Y" ]]; then
     set -e
 fi
 # ---------- CONFIGURATION ----------
@@ -14,19 +14,22 @@ AWS_CRED_FILE="$HOME/.aws/credentials"
 
 # ---------- STEP 1: Validate Terraform ----------
 echo "---------- STEP 1: Validating Terraform ----------"
-cd "$ENV_DIR" 
-
+if [ ! -d "$ENV_DIR" ]; then
+    echo "ERROR: Environment directory not found!"
+    exit 1
+fi
+cd "$ENV_DIR"
 terraform init
 terraform validate
 
 # ---------- STEP 2: Destroy Terraform resources safely ----------
 echo "---------- STEP 2: Destroying Terraform Resources ----------"
 terraform destroy --auto-approve
-echo -e "\n"
+echo -e " "
 
 # -------- STEP 3: Prompt for TF state bucket destruction --------
 read -p "Do you want to destroy the Terraform state bucket? (Y/N): " DESTROY_BUCKET
-if [[ "$DESTROY_BUCKET" == "Y" ]]; then
+if [[ "${DESTROY_BUCKET^^}" == "Y" ]]; then
     echo "---------- STEP 3: Destroying Terraform state bucket ----------"
     cd "$PROJECT_TFSTATE_PATH" 
 
@@ -56,7 +59,7 @@ fi
 # ---------- STEP 5: Clean up AWS credentials ----------
 if [ -f "$AWS_CRED_FILE" ]; then
     read -p "Do you want to remove AWS credentials? (Y/N): " RM_CRED
-    if [[ "$RM_CRED" == "Y" ]]; then
+    if [[ "${RM_CRED^^}" == "Y" ]]; then
         echo "---------- STEP 5: Removing AWS credentials ----------"
         rm -f "$AWS_CRED_FILE"
     fi
